requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-openai
langchain-groq
langchain-ollama
python-dotenv
requests
loguru
qrcode
pycloudflared
pytest
pytest-asyncio
ruff
pydantic
python-telegram-bot
pycaw
comtypes
psutil
moondream
httpx
pyautogui
pygetwindow
pillow
pynput
python-docx
pandas
PyPDF2
openpyxl
python-pptx
tictactoe-py
python-chess
pywin32
playwright
beautifulsoup4
googlesearch-python
opencv-python
pyfiglet
rich
thefuzz
python-Levenshtein
watchdog
typer
pyperclip